refactor(sync): ProviderSync interface + shared host-git ports (Phase 7.1)
Contracts-only, zero behaviour change — the spine the docker/cloud facades
(7.2–7.4) delegate to.

- core/src/sync/provider-sync.ts: the GROUPED `ProviderSync` interface (every
  shared sync op named once: resyncWorkspace/seedAgentConfig/seedCredentials/
  extractCredentials/seedGitIdentity/seedEnvFiles/applyCarry) + `CarryApplyResult`
  (mirrors docker's CopyCarryResult shape).
- Hoist the pure `SyncContext` interface into core/src/sync/context.ts (a Tier-1
  contract, so the facade can name it); sandbox-core re-exports it so existing
  importers are unchanged. `makeSyncContext` stays in sandbox-core.
- Provider gains `sync?(box): ProviderSync`; widen `resyncWorkspace?(box, onLog?)`
  so log streaming survives the facade indirection.
- Factor the 5 host-side resync ports out of docker's makeDockerResyncPorts into
  a shared `makeHostGitPorts()` in the sandbox-core git concern (host git is
  provider-neutral). Docker's ports = { ...makeHostGitPorts(), boxGit,
  probeUntrackedTokens, applyTarToBox } — byte-identical; the reuse point cloud
  resync (7.5) needs.

Gate: build + typecheck + lint green; core/sandbox-core/sandbox-docker tests
green (git-resync golden unchanged).
