+++ packages/sandbox-core/src/sync/context.ts
@@ -3,27 +3,17 @@
  * or its per-call plan. Assembled once by the provider (docker/cloud) at create
  * or session-start and threaded through every concern, so concern signatures
  * stay `(ctx, transport, plan?)`.
+ *
+ * The interface itself is a Tier-1 contract that lives in `@agentbox/core` (so
+ * the `ProviderSync` facade there can also name it); it's re-exported here so
+ * existing `@agentbox/sandbox-core` importers are unchanged. The builder
+ * `makeSyncContext` stays here — it defaults `hostHome` to the OS home dir.
  */
 
 import { homedir } from 'node:os';
+import type { SyncContext } from '@agentbox/core';
 
-export interface SyncContext {
-  /** Friendly box name (`agentbox/<name>` branch, tmux session, logs). */
-  boxName: string;
-  /** Stable box id. */
-  boxId: string;
-  /** Which provider assembled this context. */
-  provider: 'docker' | 'cloud';
-  /** Absolute host workspace dir mounted at `boxWorkspace`. */
-  hostWorkspace: string;
-  /** Project root (nearest `agentbox.yaml` ancestor, else `hostWorkspace`). */
-  projectRoot: string;
-  /** In-box workspace mount. Always `/workspace` today; overridable for tests. */
-  boxWorkspace: string;
-  /** Host home dir (source of `~/.claude` etc.). Overridable for tests. */
-  hostHome: string;
-  onLog: (line: string) => void;
-}
+export type { SyncContext } from '@agentbox/core';
 
 export interface SyncContextInit {
   boxName: string;
